backend/pyproject.toml
[project]
name = "recall"
version = "0.3.2"
description = "Local-first system to store, recall, and semantically search personal documents"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.60.0",
    "groq>=1.0.0",
    "logging>=0.4.9.6",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "sentence-transformers>=5.2.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.40.0",
]

[tool.setuptools]
packages = ["app"]

[tool.ruff]
select = ["F", "E9"]
ignore = ["E", "W", "I", "N", "D", "UP", "ANN"]

[tool.pyright]
typeCheckingMode = "basic"
